Move pkg/util/iptables to staged k8s.io/kube-proxy repo
(The only non-kube-proxy user is kubelet, and it only uses it to help
out kube-proxy.)
