+++ pkg/kubelet/kubelet_network_linux.go
@@ -23,7 +23,7 @@ import (
 
 	"k8s.io/apimachinery/pkg/util/wait"
 	"k8s.io/klog/v2"
-	utiliptables "k8s.io/kubernetes/pkg/util/iptables"
+	utiliptables "k8s.io/kube-proxy/pkg/util/iptables"
 )
 
 const (
