+++ staging/publishing/import-restrictions.yaml
@@ -212,9 +212,11 @@
 
 - baseImportPath: "./staging/src/k8s.io/kube-proxy"
   allowedImports:
+  - k8s.io/api
   - k8s.io/apimachinery
   - k8s.io/component-base
   - k8s.io/klog
+  - k8s.io/kube-proxy
   - k8s.io/utils
 
 - baseImportPath: "./staging/src/k8s.io/kube-scheduler"
