+++ staging/src/k8s.io/kube-proxy/pkg/util/iptables/iptables.go
@@ -16,6 +16,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
+//nolint:logcheck // Mostly-deprecated API, not worth fixing
 package iptables
 
 import (
