+++ staging/src/k8s.io/kube-proxy/pkg/util/iptables/testing/fake_test.go
@@ -25,7 +25,7 @@ import (
 
 	"github.com/lithammer/dedent"
 
-	"k8s.io/kubernetes/pkg/util/iptables"
+	"k8s.io/kube-proxy/pkg/util/iptables"
 )
 
 func TestFakeIPTables(t *testing.T) {
