+++ staging/src/k8s.io/kube-proxy/pkg/util/iptables/testing/parse_test.go
@@ -26,7 +26,7 @@ import (
 
 	"github.com/lithammer/dedent"
 
-	"k8s.io/kubernetes/pkg/util/iptables"
+	"k8s.io/kube-proxy/pkg/util/iptables"
 	"k8s.io/utils/ptr"
 )
